refactor component structure and import paths; add LoadMore, TotalCourse, and TopInstructors components
