+++ src/app/(root)/page.jsx
@@ -1,10 +1,10 @@
-import BecomeInstructor from "../components/BecomeInstructor";
-import TopInstructors from "../components/TopInstructors";
+import BecomeInstructor from "../../components/home/BecomeInstructor";
+import TopInstructors from "../../components/home/TopInstructors";
 
 export default function Home() {
   return (
     <div className="flex flex-col">
-      
+
     </div>
   );
 }
