+++ src/app/layout.jsx
@@ -1,7 +1,7 @@
 import { Jost } from "next/font/google";
 import "./globals.css";
-import Navbar from "./components/Navbar";
-import Footer from "./components/Footer";
+import Navbar from "../components/shared/Navbar";
+import Footer from "../components/shared/Footer";
 
 
 const jost = Jost({
