+++ src/app/page.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import TotalCourse from './components/TotalCourse'
-import CourseSubjects from './components/CourseSubjects'
-import TopInstructors from './components/TopInstructors'
-import BecomeInstructor from './components/BecomeInstructor'
+import TotalCourse from '../components/home/TotalCourse'
+import CourseSubjects from '../components/home/CourseSubjects'
+import TopInstructors from '../components/home/TopInstructors'
+import BecomeInstructor from '../components/home/BecomeInstructor'
 
 export default function Home() {
   return (
